add features
- Match typst files in subdirectories
- Add a way to get the typst files from a zip file
    - It extracts them into a temporary directory
    - Mostly to make it very easy to use the tool with a downloaded
archive from typst.app
- Add a cli interface
    - Option to ignore certain decks
    - Option to decide where to find the typst files
- Add a check for duplicate ids option: to do so, I had to separate the
parsing of the typst files from the creation of the anki cards
- Add progress bar
    - Big difficulty and design choice: I wanted to still be able to
interrupt the process while progress bars were displaying, but ctrl+c
can't be caught in the middle of a progress bar, because of the
rendering of the progress bar. **I haven't found a better solution, but
my current implementation isn't really clean. Improvements are
welcome.**
    - Wanted to keep the project zero-dependency, so I had to implement
my own progress bar; Maybe we could eventually allow a dependency for a
better progress bar implementation
- Add a dry-run mode, mostly for testing